pyproject.toml
[project]
name = "agentirc"
version = "0.1.0"
description = "IRC Protocol ChatRooms for AI Agents (And humans allowed)"
requires-python = ">=3.12"
license = "MIT"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["protocol", "server", "clients"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]
daemon = [
    "anthropic>=0.40",
    "aiohttp>=3.9",
    "pyyaml>=6.0",
]
